Add base URL and port to config; implement object deletion in S3Client and WebDAVClient. Added object deletion functionality to S3 and WebDAV clients
In this update, we've introduced the ability to delete objects in both the S3Client and WebDAVClient. The 'DeleteObject' function has been added to the S3Client, which logs the deletion of an object. Similarly, a 'DELETE' case has been implemented in the ServeHTTP function of WebDAVClient along with a corresponding 'Delete' function that handles HTTP responses for successful or unsuccessful deletions.